# pozoristeback/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teatar Maska API",
        "description": "API dokumentacija za seminarski rad",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/izvodjenja": {
            "get": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Prikaz svih planiranih izvođenja",
                "operationId": "abf404568c68dc3dd5ef113178385b7f",
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                }
            },
            "post": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Kreiranje novog izvođenja i automatsko generisanje karata",
                "description": "Kreira izvođenje i na osnovu kapaciteta sale generiše karte (npr. A1, A2...).",
                "operationId": "d58a3b6639affa0053755667ee1aae41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "predstava_id",
                                    "sala_id",
                                    "datum_izvodjenja",
                                    "vreme_pocetka",
                                    "osnovna_cena"
                                ],
                                "properties": {
                                    "predstava_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sala_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "datum_izvodjenja": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-20"
                                    },
                                    "vreme_pocetka": {
                                        "type": "string",
                                        "example": "20:00"
                                    },
                                    "osnovna_cena": {
                                        "type": "number",
                                        "example": 1200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Izvođenje kreirano i karte generisane"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/izvodjenja/{id}": {
            "get": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Detalji jednog izvođenja",
                "operationId": "04274c3d548833bd175b1acef8d0a04e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                }
            },
            "put": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Ažuriranje podataka o izvođenju",
                "operationId": "24a7c23d5439bd4e3d31da051699620a",
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Brisanje izvođenja i svih njegovih karata",
                "description": "Briše izvođenje. Zbog 'cascadeOnDelete' u migraciji, \nautomatski će se obrisati i sve karte vezane za ovo izvođenje.",
                "operationId": "a8d1f51281a274a67dde746f9a5ae463",
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/predstave/{predstavaId}/izvodjenja": {
            "get": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Sva izvođenja određene predstave",
                "description": "Vraća sva izvođenja za određenu predstavu (filtriranje).",
                "operationId": "f3b0f7ae56419cbb1ecdd9dbd22e4870",
                "parameters": [
                    {
                        "name": "predstavaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                }
            }
        },
        "/api/sale/{salaId}/izvodjenja": {
            "get": {
                "tags": [
                    "Izvođenja"
                ],
                "summary": "Prikaz svih izvođenja u određenoj sali",
                "description": "Vraća sva izvođenja u određenoj sali.",
                "operationId": "820fc54a460d9c2330faede641047b96",
                "parameters": [
                    {
                        "name": "salaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    }
                }
            }
        },
...
        "/api/predstave/{id}/izvodjenja": {
            "get": {
                "tags": [
                    "Predstave"
                ],
                "summary": "Prikaz svih izvođenja za određenu predstavu",
                "operationId": "7c2a31d9511f3fc632866ce97061b7a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID predstave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno učitana izvođenja"
                    },
                    "404": {
                        "description": "Predstava nije pronađena"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesite vaš API token koji ste dobili pri prijavi (login) ili registraciji.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}